+++ jupiter-example/src/main/java/org/jupiter/example/round/SyncJupiterClient.java
@@ -24,7 +24,6 @@ import org.jupiter.rpc.JClient;
 import org.jupiter.rpc.consumer.ProxyFactory;
 import org.jupiter.serialization.SerializerType;
 import org.jupiter.transport.JConnector;
-import org.jupiter.transport.JOption;
 import org.jupiter.transport.exception.ConnectFailedException;
 import org.jupiter.transport.netty.JNettyTcpConnector;
 
@@ -38,11 +37,12 @@ public class SyncJupiterClient {
 
     static {
         SystemPropertyUtil.setProperty("jupiter.transport.codec.low_copy", "true");
+        SystemPropertyUtil.setProperty("io.netty.allocator.type", "pooled");
+        SystemPropertyUtil.setProperty("io.netty.noPreferDirect", "true");
     }
 
     public static void main(String[] args) {
         JClient client = new DefaultClient().withConnector(new JNettyTcpConnector());
-        client.connector().config().setOption(JOption.PREFER_DIRECT, false);
 
         // 连接RegistryServer
         client.connectToRegistryServer("127.0.0.1:20001");