+++ jupiter-transport/jupiter-transport-api/src/main/java/org/jupiter/transport/JOption.java
@@ -158,16 +158,6 @@ public final class JOption<T> extends AbstractConstant<JOption<T>> {
     public static final JOption<Boolean> ALLOW_HALF_CLOSURE = valueOf("ALLOW_HALF_CLOSURE");
 
     /**
-     * 是否使用 direct buffer.
-     */
-    public static final JOption<Boolean> PREFER_DIRECT = valueOf("PREFER_DIRECT");
-
-    /**
-     * Netty的选项, 是否启用pooled buf allocator.
-     */
-    public static final JOption<Boolean> USE_POOLED_ALLOCATOR = valueOf("USE_POOLED_ALLOCATOR");
-
-    /**
      * Netty的选项, write高水位线.
      */
     public static final JOption<Integer> WRITE_BUFFER_HIGH_WATER_MARK = valueOf("WRITE_BUFFER_HIGH_WATER_MARK");
@@ -200,8 +190,6 @@ public final class JOption<T> extends AbstractConstant<JOption<T>> {
         options.add(SO_BACKLOG);
         options.add(IP_TOS);
         options.add(ALLOW_HALF_CLOSURE);
-        options.add(PREFER_DIRECT);
-        options.add(USE_POOLED_ALLOCATOR);
         options.add(WRITE_BUFFER_HIGH_WATER_MARK);
         options.add(WRITE_BUFFER_LOW_WATER_MARK);
         options.add(IO_RATIO);